+++ lib/twilio-ruby/rest/video.rb
@@ -29,6 +29,15 @@ module Twilio
       end
 
       ##
+      # @param [String] sid `CJxx…xx` A system-generated 34-character string that
+      #   uniquely identifies this Composition.
+      # @return [Twilio::REST::Video::V1::CompositionInstance] if sid was passed.
+      # @return [Twilio::REST::Video::V1::CompositionList]
+      def compositions(sid=:unset)
+        self.v1.compositions(sid)
+      end
+
+      ##
       # @return [Twilio::REST::Video::V1::CompositionSettingsInstance]
       def composition_settings
         self.v1.composition_settings()
@@ -47,15 +56,6 @@ module Twilio
       # @return [Twilio::REST::Video::V1::RecordingSettingsInstance]
       def recording_settings
         self.v1.recording_settings()
-      end
-
-      ##
-      # @param [String] sid `CJxx…xx` A system-generated 34-character string that
-      #   uniquely identifies this Composition.
-      # @return [Twilio::REST::Video::V1::CompositionInstance] if sid was passed.
-      # @return [Twilio::REST::Video::V1::CompositionList]
-      def compositions(sid=:unset)
-        self.v1.compositions(sid)
       end
 
       ##