+++ spec/integration/api/v2010/account/sip/ip_access_control_list/ip_address_spec.rb
@@ -39,21 +39,19 @@ describe 'IpAddress' do
                   "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "date_created": "Mon, 20 Jul 2015 17:27:10 +0000",
                   "date_updated": "Mon, 20 Jul 2015 17:27:10 +0000",
-                  "friendly_name": "aaa",
+                  "friendly_name": "friendly_name",
                   "ip_access_control_list_sid": "ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                  "ip_address": "192.1.1.2",
+                  "ip_address": "ip_address",
+                  "cidr_prefix_length": 32,
                   "sid": "IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses/IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json"
               }
           ],
-          "last_page_uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses.json?PageSize=50&Page=0",
           "next_page_uri": null,
-          "num_pages": 1,
           "page": 0,
           "page_size": 50,
           "previous_page_uri": null,
           "start": 0,
-          "total": 1,
           "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses.json?PageSize=50&Page=0"
       }
       ]
@@ -75,14 +73,11 @@ describe 'IpAddress' do
           "end": 0,
           "first_page_uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses.json?PageSize=50&Page=0",
           "ip_addresses": [],
-          "last_page_uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses.json?PageSize=50&Page=0",
           "next_page_uri": null,
-          "num_pages": 1,
           "page": 0,
           "page_size": 50,
           "previous_page_uri": null,
           "start": 0,
-          "total": 1,
           "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses.json?PageSize=50&Page=0"
       }
       ]
@@ -123,9 +118,10 @@ describe 'IpAddress' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "Mon, 20 Jul 2015 17:27:10 +0000",
           "date_updated": "Mon, 20 Jul 2015 17:27:10 +0000",
-          "friendly_name": "aaa",
+          "friendly_name": "friendly_name",
           "ip_access_control_list_sid": "ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "ip_address": "192.1.1.2",
+          "ip_address": "ip_address",
+          "cidr_prefix_length": 32,
           "sid": "IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses/IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json"
       }
@@ -166,9 +162,10 @@ describe 'IpAddress' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "Mon, 20 Jul 2015 17:27:10 +0000",
           "date_updated": "Mon, 20 Jul 2015 17:27:10 +0000",
-          "friendly_name": "aaa",
+          "friendly_name": "friendly_name",
           "ip_access_control_list_sid": "ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "ip_address": "192.1.1.2",
+          "ip_address": "ip_address",
+          "cidr_prefix_length": 32,
           "sid": "IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses/IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json"
       }
@@ -209,9 +206,10 @@ describe 'IpAddress' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "Mon, 20 Jul 2015 17:27:10 +0000",
           "date_updated": "Mon, 20 Jul 2015 17:27:10 +0000",
-          "friendly_name": "aaa",
+          "friendly_name": "friendly_name",
           "ip_access_control_list_sid": "ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "ip_address": "192.1.1.2",
+          "ip_address": "ip_address",
+          "cidr_prefix_length": 32,
           "sid": "IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "uri": "/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/SIP/IpAccessControlLists/ALaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/IpAddresses/IPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json"
       }